interview-coach/requirements.txt
# ============================================================================
# AI 模拟面试系统 - Python 依赖
# ============================================================================
# 安装方式: pip install -r requirements.txt
# Python 版本要求: >=3.9

# ============================================================================
# LLM API 客户端
# ============================================================================

# OpenAI 官方库（兼容所有 OpenAI API 格式的服务）
openai>=1.0.0

# ============================================================================
# 文档处理
# ============================================================================

# PDF 处理
pymupdf>=1.23.0                        # 强大的 PDF 解析（推荐）
PyPDF2>=3.0.0                          # 备用 PDF 解析

# ============================================================================
# Web 搜索工具
# ============================================================================

ddgs>=9.0.0                            # DuckDuckGo 搜索
requests>=2.31.0                       # HTTP 请求

# ============================================================================
# Web UI 界面
# ============================================================================

gradio>=4.0.0                          # Web UI 框架

# ============================================================================
# 环境配置和工具
# ============================================================================

python-dotenv>=1.0.0                   # 环境变量管理
loguru>=0.7.0                          # 日志库

# ============================================================================
# 可选依赖
# ============================================================================

# 测试工具（可选）
# pytest>=8.0.0
# pytest-asyncio>=0.23.0

# 代码质量工具（可选）
# black>=24.0.0
# isort>=5.13.0
# flake8>=7.0.0
